Java app for Assignment 2
For assignment 2, I reused this Java library:  https://github.com/Hoofnuts/SENG560_Project1/tree/main  to make a command-line calculator app in Java.  Intent would be to expand this calculator to eventually have an interface. Given the assignment tis focused on reusing a library, I opted to forgo actually building the graphical interface and instead focus on the library integration.